# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/auth/register":{"post":{"operationId":"AuthController_register","summary":"User register to platform","description":"## User register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"},"examples":{"user_1":{"value":{"firstName":"John","lastName":"Doe","email":"johndoe@example.com","password":"1232@asdS"}}}}}},"responses":{"201":{"description":"User created successfully!","content":{"application/json":{"examples":{"created_user":{"summary":"Response after register","value":{"access_token":"eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjQ0MWNkNmJlMWQ0ZTBiNDRjNzA3NDk2IiwiaWF0IjoxNjgyMDM0MDI3LCJleHAiOjE2ODIwMzc2Mjd9.AH4z7uDWuEDjOs8sesB0ItxKUJ2M3rjul1D1mmjAKieOZblej5mp0JQE5IdgB9LlAOzOtKOLL5RWhxLCZ-YskvoRA7Yqza_rOjfIHeNseC3M66kKYqORN07aZDiA2OWhT3pXBqoKRCUBQCKLgMCAPT-CHryc0wUQGaKxP8YJO8dwIhGtjADchmzNJVBs4G7qYnpZAsORayd5GNfgoLpWmVFIBHSnPLNIL4dL8dLof0GBmVhdjhnHIUXYQlqL1wiwsmxmUC9TU2uiChm-TAhuiQyVwFokSySBJzBrLmEtgy89aaR0YizFK-QMg2xW3cJiiRzEBigTdsR0kvdUlk5GOg","refresh_token":"eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjQ0MWNkNmJlMWQ0ZTBiNDRjNzA3NDk2IiwiaWF0IjoxNjgyMDM0MDI3LCJleHAiOjE2ODIwNTkyMjd9.aKNZymKdf3VEbPkda2cYYTS7KlpCbTqdXP30LREQ2b_fJ8q8cA0OyNEARK3Jm5yGsKoNd3txi54XmEbf19LC9CuDf9kwgLasPizEeMZsAJqSbSguzE4-9b4sSdf22GyipCcZJpkXkp01Bew04J8Y4FqhNARONsWzySXg8_VVWOGkfHGJVHFs7xYyVvmt3RErJwRM5s1Ou1ok7VW62FSTSAvXw6-qsHp5T7kXo73jECBqSuNEs5JcdluoBjdaAxggHYaDgTXoRh7y4Mn_fVKCQarAsUAxg6w0fxc8Gj0nP1ct3-GjG-Of-0O-iF7uynI2Lnq_On7WUsH7rFSysNyHUg"}}}}}},"400":{"description":"Validation failed","content":{"application/json":{"examples":{"invalid_email_password":{"value":{"statusCode":400,"message":["Email must be an email","Password is not strong enough"],"error":"Bad Request"}},"some_fields_missing":{"value":{"statusCode":400,"message":["lastName must be shorter than or equal to 50 characters","lastName should not be empty"],"error":"Bad Request"}}}}}},"409":{"description":"Conflict user info","content":{"application/json":{"examples":{"email_duplication":{"value":{"statusCode":409,"message":"Email already existed!!","error":"Conflict"}}}}}}},"tags":["Authorize"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"User login to platform","description":"## User login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"},"examples":{"user_1":{"value":{"email":"johndoe@example.com","password":"1232@asdS"}}}}}},"responses":{"201":{"description":""},"401":{"description":"Unauthorized","content":{"application/json":{"example":{"statusCode":400,"message":"Wrong credentials!!","error":"Bad Request"}}}}},"tags":["Authorize"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refreshAccessToken","summary":"Refresh Access-token for user","description":"## Refresh Access-token","parameters":[],"responses":{"201":{"description":""}},"tags":["Authorize"]}},"/users":{"post":{"operationId":"UsersController_create","summary":"Admin create new user","description":"\n* Only admin can use this API\n\n* Admin create user and give some specific information","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"get":{"operationId":"UsersController_findAll","summary":"Admin find user list","description":"# Admin find user list","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","summary":"Admin find user by ID","description":"# Admin find user ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"patch":{"operationId":"UsersController_update","summary":"Admin update user details","description":"# Admin update user details","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","summary":"Admin delete user details","description":"# Admin delete user details","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Users"]}},"/posts":{"post":{"operationId":"PostsController_create","summary":"User share new post","description":"# User share new post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}}}},"tags":["Posts"]},"get":{"operationId":"PostsController_findAll","summary":"Get all posts","description":"# Get all posts","parameters":[{"name":"offset","required":true,"in":"query","examples":{"0":{"value":0,"description":"Start from 0"},"10":{"value":10,"description":"Skip 10 collection"}},"schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","examples":{"10":{"value":10,"description":"Get 10 collection"},"50":{"value":50,"description":"Get 50 collection"}},"schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Posts"]}},"/posts/{id}":{"get":{"operationId":"PostsController_findOne","summary":"Get post by ID","description":"# Get post by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}}}},"tags":["Posts"]},"patch":{"operationId":"PostsController_update","summary":"Update post details","description":"# Update post details","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}}}},"tags":["Posts"]},"delete":{"operationId":"PostsController_remove","summary":"Delete post","description":"# Delete posts","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Posts"]}}},"info":{"title":"Youtube Sharing App","description":"## Youtube Sharing App API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"RegisterDto":{"type":"object","properties":{"firstName":{"type":"string","maxLength":50},"lastName":{"type":"string","maxLength":50},"email":{"type":"string","maxLength":50},"password":{"type":"string"}},"required":["firstName","lastName","email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string","maxLength":50},"lastName":{"type":"string","maxLength":50},"email":{"type":"string","maxLength":50},"password":{"type":"string"}},"required":["firstName","lastName","email","password"]},"User":{"type":"object","properties":{"friendlyId":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"current_refresh_token":{"type":"string"},"_id":{"type":"object"},"id":{"type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["friendlyId","firstName","lastName","email","password","current_refresh_token"]},"UpdateUserDto":{"type":"object","properties":{"firstName":{"type":"string","maxLength":50},"lastName":{"type":"string","maxLength":50}}},"CreatePostDto":{"type":"object","properties":{"title":{"type":"string"},"link":{"type":"string"},"description":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["title","link","description","user"]},"Post":{"type":"object","properties":{"title":{"type":"string"},"link":{"type":"string"},"description":{"type":"string"},"like":{"type":"array","items":{"type":"string"}},"unlike":{"type":"array","items":{"type":"string"}},"user":{"$ref":"#/components/schemas/User"},"_id":{"type":"object"},"id":{"type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["title","link","like","unlike","user"]},"UpdatePostDto":{"type":"object","properties":{"like":{"type":"array","items":{"type":"string"}},"unlike":{"type":"array","items":{"type":"string"}}},"required":["like","unlike"]}}}}